+++ Sources/InfluxDBSwift/WriteAPI.swift
@@ -11,7 +11,7 @@ import FoundationNetworking
 #endif
 import Gzip
 
-/// The asynchronous API to Write time-series data into InfluxDB 2.0.
+/// The asynchronous API to Write time-series data into InfluxDB 2.x.
 ///
 /// ### Example: ###
 /// ````
@@ -104,7 +104,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     ///   - completion: completion handler to receive the data and the error objects
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -131,7 +131,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     ///   - completion: handler to receive the data and the error objects
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -159,7 +159,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     ///   - completion: completion handler to receive the data and the error objects
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -186,7 +186,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     ///   - completion: handler to receive the data and the error objects
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -214,7 +214,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     ///   - completion: completion handler to receive the data and the error objects
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -241,7 +241,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     ///   - completion: handler to receive the data and the error objects
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -270,7 +270,7 @@ public class WriteAPI {
     ///   - completion: completion handler to receive the `Swift.Result`
     ///
     /// - SeeAlso: `Swift.Result`
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -298,7 +298,7 @@ public class WriteAPI {
     ///   - completion: completion handler to receive the `Swift.Result`
     ///
     /// - SeeAlso: `Swift.Result`
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -327,7 +327,7 @@ public class WriteAPI {
     ///   - completion: completion handler to receive the `Swift.Result`
     ///
     /// - SeeAlso: `Swift.Result`
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -355,7 +355,7 @@ public class WriteAPI {
     ///   - completion: completion handler to receive the `Swift.Result`
     ///
     /// - SeeAlso: `Swift.Result`
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -384,7 +384,7 @@ public class WriteAPI {
     ///   - completion: completion handler to receive the `Swift.Result`
     ///
     /// - SeeAlso: `Swift.Result`
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -412,7 +412,7 @@ public class WriteAPI {
     ///   - completion: completion handler to receive the `Swift.Result`
     ///
     /// - SeeAlso: `Swift.Result`
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     public func write(bucket: String? = nil,
                       org: String? = nil,
                       precision: InfluxDBClient.TimestampPrecision? = nil,
@@ -441,7 +441,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     /// - Returns: Publisher to attach a subscriber
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
     public func write(bucket: String? = nil,
                       org: String? = nil,
@@ -466,7 +466,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     /// - Returns: Publisher to attach a subscriber
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
     public func write(bucket: String? = nil,
                       org: String? = nil,
@@ -496,7 +496,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     /// - Returns: Publisher to attach a subscriber
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
     public func write(bucket: String? = nil,
                       org: String? = nil,
@@ -521,7 +521,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     /// - Returns: Publisher to attach a subscriber
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
     public func write(bucket: String? = nil,
                       org: String? = nil,
@@ -551,7 +551,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     /// - Returns: Publisher to attach a subscriber
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
     public func write(bucket: String? = nil,
                       org: String? = nil,
@@ -576,7 +576,7 @@ public class WriteAPI {
     ///   - responseQueue: The queue on which api response is dispatched.
     /// - Returns: Publisher to attach a subscriber
     ///
-    /// - SeeAlso: https://docs.influxdata.com/influxdb/v2.0/reference/syntax/line-protocol/
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
     @available(OSX 10.15, iOS 13.0, tvOS 13.0, watchOS 6.0, *)
     public func write(bucket: String? = nil,
                       org: String? = nil,
@@ -594,6 +594,164 @@ public class WriteAPI {
                 }
             }
         }.eraseToAnyPublisher()
+    }
+    #endif
+
+    #if swift(>=5.5)
+    /// Write time-series data asynchronously into InfluxDB.
+    ///
+    /// - Parameters:
+    ///   - bucket:  The destination bucket for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - org: The destination organization for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - precision: The precision for the unix timestamps within the body line-protocol.
+    ///   - record: The record to write.
+    ///   - responseQueue: The queue on which api response is dispatched.
+    ///
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    public func write(bucket: String? = nil,
+                      org: String? = nil,
+                      precision: InfluxDBClient.TimestampPrecision? = nil,
+                      record: String,
+                      responseQueue: DispatchQueue = .main) async throws {
+        try await self.write(
+                bucket: bucket,
+                org: org,
+                precision: precision,
+                records: [record],
+                responseQueue: responseQueue)
+    }
+
+    /// Write time-series data asynchronously into InfluxDB.
+    ///
+    /// - Parameters:
+    ///   - bucket:  The destination bucket for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - org: The destination organization for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - precision: The precision for the unix timestamps within the body line-protocol.
+    ///   - records: The records to write.
+    ///   - responseQueue: The queue on which api response is dispatched.
+    ///
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    public func write(bucket: String? = nil,
+                      org: String? = nil,
+                      precision: InfluxDBClient.TimestampPrecision? = nil,
+                      records: [String],
+                      responseQueue: DispatchQueue = .main) async throws {
+        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) -> Void in
+            self.postWrite(bucket, org, precision, records, responseQueue) { result -> Void in
+                switch result {
+                case .success:
+                    continuation.resume(returning: ())
+                case let .failure(error):
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+
+    /// Write time-series data asynchronously into InfluxDB.
+    ///
+    /// - Parameters:
+    ///   - bucket:  The destination bucket for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - org: The destination organization for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - precision: The precision for the unix timestamps within the body line-protocol.
+    ///   - point: The `Point` to write.
+    ///   - responseQueue: The queue on which api response is dispatched.
+    ///
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    public func write(bucket: String? = nil,
+                      org: String? = nil,
+                      precision: InfluxDBClient.TimestampPrecision? = nil,
+                      point: InfluxDBClient.Point,
+                      responseQueue: DispatchQueue = .main) async throws {
+        try await self.write(
+                bucket: bucket,
+                org: org,
+                precision: precision,
+                points: [point],
+                responseQueue: responseQueue)
+    }
+
+    /// Write time-series data asynchronously into InfluxDB.
+    ///
+    /// - Parameters:
+    ///   - bucket:  The destination bucket for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - org: The destination organization for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - precision: The precision for the unix timestamps within the body line-protocol.
+    ///   - points: The `Points` to write.
+    ///   - responseQueue: The queue on which api response is dispatched.
+    ///
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    public func write(bucket: String? = nil,
+                      org: String? = nil,
+                      precision: InfluxDBClient.TimestampPrecision? = nil,
+                      points: [InfluxDBClient.Point],
+                      responseQueue: DispatchQueue = .main) async throws {
+        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) -> Void in
+            self.postWrite(bucket, org, precision, points, responseQueue) { result -> Void in
+                switch result {
+                case .success:
+                    continuation.resume(returning: ())
+                case let .failure(error):
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
+    }
+
+    /// Write time-series data asynchronously into InfluxDB.
+    ///
+    /// - Parameters:
+    ///   - bucket:  The destination bucket for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - org: The destination organization for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - precision: The precision for the unix timestamps within the body line-protocol.
+    ///   - tuple: The `Tuple` to write.
+    ///   - responseQueue: The queue on which api response is dispatched.
+    ///
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    public func write(bucket: String? = nil,
+                      org: String? = nil,
+                      precision: InfluxDBClient.TimestampPrecision? = nil,
+                      tuple: InfluxDBClient.Point.Tuple,
+                      responseQueue: DispatchQueue = .main) async throws {
+        try await self.write(
+                bucket: bucket,
+                org: org,
+                precision: precision,
+                tuples: [tuple],
+                responseQueue: responseQueue)
+    }
+
+    /// Write time-series data asynchronously into InfluxDB.
+    ///
+    /// - Parameters:
+    ///   - bucket:  The destination bucket for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - org: The destination organization for writes. Takes either the `ID` or `Name` interchangeably.
+    ///   - precision: The precision for the unix timestamps within the body line-protocol.
+    ///   - tuples: The `Tuples` to write.
+    ///   - responseQueue: The queue on which api response is dispatched.
+    ///
+    /// - SeeAlso: https://docs.influxdata.com/influxdb/latest/reference/syntax/line-protocol/
+    @available(macOS 10.15, iOS 13.0, watchOS 6.0, tvOS 13.0, *)
+    public func write(bucket: String? = nil,
+                      org: String? = nil,
+                      precision: InfluxDBClient.TimestampPrecision? = nil,
+                      tuples: [InfluxDBClient.Point.Tuple],
+                      responseQueue: DispatchQueue = .main) async throws {
+        try await withCheckedThrowingContinuation { (continuation: CheckedContinuation<Void, Error>) -> Void in
+            postWrite(bucket, org, precision, tuples, responseQueue) { result -> Void in
+                switch result {
+                case .success:
+                    continuation.resume(returning: ())
+                case let .failure(error):
+                    continuation.resume(throwing: error)
+                }
+            }
+        }
     }
     #endif